Change Encoder to only write when asked
Encoder now only writes when the Write method is called. Due to this
simplification it no longer needs to store an io.Writer and so this is
passed to the Write method.

There is also a WriteAndPad method so any padding that needs to be done
is explicit.

Added some documentation